fix(driver): rust-analyzer is confused by Extra
