+++ compio-driver/src/sys/iocp/mod.rs
@@ -28,7 +28,7 @@ mod wait;
 
 /// Extra data attached for IOCP.
 #[repr(C)]
-pub struct Extra {
+pub(in crate::sys) struct Extra {
     overlapped: Overlapped,
 }
 
@@ -358,7 +358,7 @@ impl Driver {
         &self.notify.port
     }
 
-    pub fn default_extra(&self) -> Extra {
+    pub(in crate::sys) fn default_extra(&self) -> Extra {
         Extra::new(self.port().as_raw_handle() as _)
     }
