+++ compio-driver/src/sys/iour/mod.rs
@@ -43,7 +43,7 @@ use crate::{
 };
 
 mod extra;
-pub use extra::Extra;
+pub(in crate::sys) use extra::Extra;
 pub(crate) mod op;
 pub(crate) use op::take_buffer;
 
@@ -318,7 +318,7 @@ impl Driver {
         has_entry
     }
 
-    pub fn default_extra(&self) -> Extra {
+    pub(in crate::sys) fn default_extra(&self) -> Extra {
         Extra::new()
     }
